Conflicts:
	common/src/main/java/com/winhxd/b2c/common/feign/store/StoreServiceClient.java
	service-projects/order-service/src/main/java/com/winhxd/b2c/order/service/impl/OrderQueryServiceImpl.java
	service-projects/store-service/src/main/java/com/winhxd/b2c/store/api/ApiOpenStoreController.java
	service-projects/store-service/src/main/java/com/winhxd/b2c/store/controller/StoreServiceController.java
	service-projects/store-service/src/main/java/com/winhxd/b2c/store/dao/StoreUserInfoMapper.java
	service-projects/store-service/src/main/java/com/winhxd/b2c/store/service/StoreService.java
	service-projects/store-service/src/main/java/com/winhxd/b2c/store/service/impl/StoreServiceImpl.java

+++ service-projects/promotion-service/src/main/java/com/winhxd/b2c/promotion/dao/CouponApplyBrandMapper.java
@@ -3,13 +3,13 @@ package com.winhxd.b2c.promotion.dao;
 import com.winhxd.b2c.common.domain.promotion.model.CouponApplyBrand;
 
 public interface CouponApplyBrandMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CouponApplyBrand record);
 
     int insertSelective(CouponApplyBrand record);
 
-    CouponApplyBrand selectByPrimaryKey(Integer id);
+    CouponApplyBrand selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CouponApplyBrand record);
