+++ service-projects/promotion-service/src/main/java/com/winhxd/b2c/promotion/dao/CouponApplyMapper.java
@@ -3,13 +3,13 @@ package com.winhxd.b2c.promotion.dao;
 import com.winhxd.b2c.common.domain.promotion.model.CouponApply;
 
 public interface CouponApplyMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CouponApply record);
 
     int insertSelective(CouponApply record);
 
-    CouponApply selectByPrimaryKey(Integer id);
+    CouponApply selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CouponApply record);
