+++ service-projects/promotion-service/src/main/java/com/winhxd/b2c/promotion/dao/CouponApplyProductMapper.java
@@ -3,13 +3,13 @@ package com.winhxd.b2c.promotion.dao;
 import com.winhxd.b2c.common.domain.promotion.model.CouponApplyProduct;
 
 public interface CouponApplyProductMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(CouponApplyProduct record);
 
     int insertSelective(CouponApplyProduct record);
 
-    CouponApplyProduct selectByPrimaryKey(Integer id);
+    CouponApplyProduct selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(CouponApplyProduct record);
